+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Espresso-Helper 0.1.2 (2018-05-05)
+
+- Remove `checkNextActivityByIntent()` because it is not needed next to `checkCurrentActivityIs`
+
+- Alter `BackgroundColorMatcher` to detect `GradientDrawable`'s default color too on API 24+
+
 # Espresso-Helper 0.1.1 (2018-05-03)
 
 - Add `checkCurrentActivityIs()` (with `ActivityLifecycleMonitorRegistry`), and `checkNextActivityByIntent()` (with `IntentsTestRule`)